New Logic for Evidence completed and tested
Fixed the logic for the several mobile devices that treated itemlists differently. So basically what would happen on some samsung devices is that an item would be tapped and it could not be tapped again unless focus was moved from the item to another item in the list. This chance was a rewrite to use buttons instead, how it should have been done right away. There are many places for optimization/code cleanup.